Move aristech stt code out of lib.rs to reside in separate transcribe.rs file.
